--- index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
-  </head>
-  <body>
-    <story-telling></story-telling>
-    <script type="module" src="./src/main2.js"></script>
-    <style>
-      @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap");
-      * {
-        font-family: "Roboto", sans-serif;
-      }
-      body {
-        padding: 0;
-        margin: 0;
-      }
-    </style>
-  </body>
-</html>
